pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sergii_Cheliuk <bormot87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
flask-httpauth = "^4.7.0"
pyjwt = "^2.6.0"
flask-sqlalchemy = "^3.0.2"
marshmallow-sqlalchemy = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
